+++ src/modules/AppointmentForm/redux/saga/index.ts
@@ -1,2 +1,2 @@
-export * from './createAppointment';
-export * from './getFormData';
+export * from './postNewAppointmentSaga';
+export * from './fetchFormDataSaga';
